[Editorial] Remove italics from loanword (ennui)
There is another instance of "ennui" earlier in the chapter that is not
italicized in the PG source. Both are italicized in the source text. It
could be the case that, instead of removing italics, this italicization
should be added to the earlier case as well. These two are the only
occurrences of "ennui" in the book, and given in the context of Ivan
having extensive French education it could be reasonable to emphasize
the "French-ness" of ennui.